// src/core/model/seo.ts
export interface Seo {
	id?: number;
	name?: string;
	title?: string;
	description?: string;
	keyWords?: string;
	imageUrl?: string;
	url?: string;
	mainUrl?: string;
	createDate?: Date;
	updateDate?: Date;
	deleteDate?: Date;
	createById?: number;
	updateById?: number;
	deleteById?: number;
	isDelete?: boolean;
}